summary

+++ src/C++/FieldTypes.h
@@ -41,7 +41,7 @@ namespace FIX
  */
 
 /// Date and Time stored as a Julian day number and number of
-/// milliseconds since midnight.  Does not perform any timezone
+/// nanoseconds since midnight.  Does not perform any timezone
 /// calculations.  All magic numbers and related calculations
 /// have been taken from:
 ///
@@ -135,7 +135,7 @@ struct DateTime
   /// Return the hour portion of the time (0-23)
   inline int getHour() const 
   {
-    return m_time / NANOS_PER_HOUR;
+    return (int)(m_time / NANOS_PER_HOUR);
   }
 
   /// Return the minute portion of the time (0-59)
